requirements.txt
Django==3.2.16
djangorestframework==3.14.0
requests==2.28.1
django_tables2==2.4.1
django_filter==22.1
django_bootstrap3==22.1
